Add agent working for CLI command
User will use CLI command using the agent
which is working as localhost server.

This is initial version for test purpose.
There are some good references for fetch API as below.

 - https://hogni.tistory.com/142
 - https://ko.javascript.info/fetch-progress

We need to consider if we will use fetch API or not.
Because IE does not support this API.